chore: fix sidebar close and open
